requirements.txt
httpx>=0.25.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0